refactor account balance to use BigDecimal for extra precision

+++ src/main/java/honours/ing/banq/info/bean/BalanceBean.java
@@ -12,7 +12,7 @@ public class BalanceBean {
     private Double balance;
 
     public BalanceBean(BankAccount bankAccount) {
-        balance = bankAccount.getBalance();
+        balance = bankAccount.getBalance().doubleValue();
     }
 
     public Double getBalance() {
